mRNA Translation
In the works! I was experimenting with the biopython library and was honestly struggling to understand it. After playing around with it for a while, I decided that I would try to make my own translate function. It seemed like a fun challenge, and honestly I just couldn't figure out the library.